backend/requirements.txt
# backend/requirements.txt
flask
flask-cors
pandas
numpy
PyJWT
bcrypt
psycopg2-binary